Dark mode (antd styles with my customizations works well):
- my custom styles that depends on dark/light mode are CSS variables
(instead of @my-variable I have to use --my-variable)

Reason:
"A major blocker for theming with color variables is using a pre-processor such as LESS/SASS. Pre-processors’ variables are static, and while pre-processing, the values of these variables get substituted in the classes where they are used.
(...)
In the generated CSS, we lost the variables and therefore cannot update the colors dynamically.
The solution: CSS properties (a.k.a. CSS variables)
On the other hand, CSS variables (which modern browsers understand) can be updated dynamically."

Links:
https://medium.com/swlh/dark-mode-using-css-variables-cf065a7fa133